apps/server/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
python-multipart==0.0.6
pydantic==2.3.0
pydantic-settings==2.0.3
sqlalchemy==2.0.20
sqlmodel==0.0.8
alembic==1.12.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.0
pytest==7.4.2
httpx==0.24.1
asyncpg==0.29.0
supabase==2.3.0
psycopg2-binary==2.9.9
uuid==1.30

# sqlalchemy

